moved all the parse tests to apfel_spec.rb
